pyproject.toml
[project]
name = "codesm"
version = "0.1.0"
description = "AI coding agent"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.40.0",
    "openai>=1.54.0",
    "ollama>=0.4.0",
    "textual>=0.90.0",
    "typer>=0.15.0",
    "pydantic>=2.10.0",
    "httpx>=0.28.0",
    "aiofiles>=24.1.0",
    "rich>=13.9.0",
    "sentence-transformers>=2.2.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]

[project.scripts]
codesm = "codesm.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["codesm"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
